requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.3
docker==5.0.3
fastapi==0.73.0
frozenlist==1.3.0
h11==0.13.0
idna==3.3
loguru==0.6.0
multidict==6.0.2
pydantic==1.9.0
python-dotenv==0.20.0
python-multipart==0.0.5
requests==2.28.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.0.1
urllib3==1.26.9
uvicorn==0.17.4
websocket-client==1.3.2
yarl==1.7.2
